refactor(pacer_free_documents): move CycleChecker to corpus_importer/utils.py
